% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ryx}
\alias{plot.ryx}
\title{ryx.plot}
\usage{
\method{plot}{ryx}(x, digits = 3)
}
\arguments{
\item{x}{list of y variable and x variables for correlation along with the data frame of correlations p values and level of significance}
}
\description{
This \code{plot} function that plots the variables against the correlation values
(data from ryx function) in a horizontal graph
}
\details{
The plot function is a wrapper for the
\href{https://ggplot2.tidyverse.org/reference/ggplot.html}{ggplot2}.
}
\examples{
x<- ryx(Boston, y="medv")
plot(x)

}
\seealso{
\link[ggplot2]{ggplot2}
}
