% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ryx}
\alias{print.ryx}
\title{Visualize ryx function results in dataframe}
\usage{
\method{print}{ryx}(x, digits = 3)
}
\arguments{
\item{x}{results from ryx function}
}
\description{
This \code{print} function presents a table of correlation values of x
variables with the specified y value(all defined from the ryx function), the corresponding
p-values and a column that specifies how significant these p-values are.
}
\details{
The print function is a wrapper for the cbind base R function
\href{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/cbind}{cbind}.
}
\examples{
x<- ryx(Boston, y="medv")
print(x)


}
