% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summarized results for ryx function in text.}
\usage{
\method{summary}{ryx}(x, digits = 3)
}
\arguments{
\item{x}{results from ryx function}
}
\description{
This \code{summary} function that summarizes the contents of the ryx function by
giving the median and range of the correlation values along with the number of values
that are significant.
}
\details{
The summary function uses the data from the results of the ryx function
}
\examples{

x<- ryx(Boston, y="medv")
summary(x)

}
