% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Visualize ryx function results in horizontal graph}
\usage{
\method{plot}{ryx}(x, digits = 3)
}
\arguments{
\item{x}{results from ryx function}
}
\description{
This \code{plot} function that plots the variables against the correlation values
(data from ryx function) in a horizontal graph
}
\details{
The plot function is a wrapper for the
\href{https://www.rdocumentation.org/packages/ggplot2/versions/3.4.0}{ggplot2}.
}
\examples{
x<- ryx(Boston, y="medv")
plot(x)

}
\seealso{
\link[ggplot2]{ggplot2}
}
