% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{ryx.print}
\usage{
ryx(data, y, x)
}
\description{
This \code{print} function presents a table of variables, correlation values of x
variables with the specified y value(all defined from the ryx function), the corresponding
p-values and a column that specifies how significant these p-values are.
}
\details{
The print function is a wrapper for the cbind base R function
\href{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/cbind}{cbind}.

#' @param x list of y variable and x variables for correlation along with the data frame of correlations p values and level of significance
}
\examples{

x<- ryx(Boston, y="medv")
print(x)


}
